requirements.txt
albumentations==1.3.0
matplotlib==3.5.2
numpy==1.23.5
Pillow==9.2.0
tqdm==4.64.1